+++ server/api/auth/googleStrategyVerifyCallback.ts
@@ -1,7 +1,8 @@
 import { Request } from 'express'
 import { Profile, VerifyFunction } from 'passport-google-oauth'
-import { UserController, UserProviderController } from '@server/controller'
 import { AuthProvider } from '@meta/user'
+import { UserController } from '@server/controller/user'
+import { UserProviderController } from '@server/controller/userProvider'
 
 export const googleStrategyVerifyCallback = async (
   req: Request,
