--- server/controller/index.ts
@@ -1,5 +1,0 @@
-export { AssessmentController } from './assessment'
-export { CycleDataController } from './cycleData'
-export { SettingsController } from './settings'
-export { UserController } from './user'
-export { UserProviderController } from './userProvider'
